Add ISOMIP+ Ocean0-2 to testing infrastructure
This merge adds the configuration files for the ISOMIP+ Ocean0,
Ocean1 and Ocean2 experiments.

The Ocean0-2 experiments default to a short forward run and have
been added to the land-ice regression tests.

Adds a template and driver script to isomip_plus for a
parameter study in which the heat and salt transfer
coefficients (GammaT and GammaS) vary over 10 values for a
forward run of the ISOMIP+ Ocean0 test case.