Add --conda_mpi flag to setup_testcase.py
Add support for a config option `use_conda_mpi` in a new
`conda` section.

With these new options, the script figures out the full path to
mpirun.

This is needed to support compass conda enviornments with mpich.
Python scripts and modules that use the netcdf4 package with
mpich support don't work properly unless they are prefixed with
`mpirun -np 1`