Enable missing_value to override default_value
This commit allows missing_value to override the default value of a field if
default_value is not specified for a field in a core's Registry.xml file. If
missing_value and default_value are both present for a field, then the field
will be initialized to default_value.

Specifying missing_value for a field in the Registry.xml will still create a
missing_value NetCDF attribute if that field is written (even if default_value
is also specified).